profiling: LCI support
Profiling is started before the communication engine, so there is
no LCI progress thread. We must therefore ensure our own progress
for LCI collective calls. All other uses of LCI collectives are
after the progress thread is enabled, so this isn't necessary
elsewhere. This is not on the critical path anyway (the collectives
are used only during profiling initialization), so it's not a
performance problem.