dist collectives: fix for propagating too much data in bcast
A flow that is part of a broadcast tree will alway propagate data
based on the amount of data it expected to receive. When the flow can
receive less data than it expects, it then propagates too much data
to its descendents in the broadcast.

The fix is to have such flows use layout x count format and to update
the count once the data arrives. If the data was allocated by an
arena, we must make sure to allocate only a single count i.e. assumes
that the arena datatype covers the maximum input size it can receive.
I don't think there are any current uses for arena datatype x count,
so this should be okay. This is necessarily the cleanest fix,
but it works for now.