lci: refactor memory registration
Memory registration used to:
  a) carry process-local data
  b) be 72 bytes long
  c) be aligned to 64 bytes

This meant that the total structure was 128 bytes which, combined with
the 40 header bytes for a get request AM, meant that get requests were
168 bytes and could not be sent as short LCI messages. By refactoring
out the core memory registration data (32 bytes) and putting it in a
union with the list item struct, we reduce the total structure size to
64 bytes. By telling the caller that the size of the mem reg is only
that of its core data, only 72 bytes need to be sent out on the wire. A
caveat is that the list item needs to be reinitialized before the mem
reg handle is returned to the object pool; it is overwritten during use.