Make the generation of binary profile files more resilient against abnormal interruptions.
The reader will only open files that contain a specific string (PARSEC_PROFILE_MAGICK) at a specific position in the header of a binary profile file.

Move the setting of that string into the header from the initialization of the file to after the file is consistent. This way, if the process is aborted and the file is only partially written, the reader will refuse to open it.