lci: many changes to how messages get processed
Many different changes involving how completed messages get moved from
the progress to the communication thread, how the comm thread processes
the completed messages, the order in which they get processed, and so
forth. This could probably be pulled into 3–4 different patches, but it
would be complex to do so with minimal benefits, so I haven't.

Includes some experimental code that has been disabled or commented out.