stats: track chain of tasks with largest exec time
Rudimentary critical path tracking, assumes that chain of tasks
with the largest cummulative execution time is the critical path.
This is approximately correct for taskpools that are strongly bound
by a critical path, but is inaccurate when there are sibling branches
that compete for the longest path: the measured ``critical path''
would appear to jump between these branches, depending on
which overtakes the other at any given point in the execution.