remote_dep: consistent calculation of max_priority
remote_deps_t::max_priority was calculated differently depending on context:
  * {parsec|dtd}_release_dep_fct ignored successors which were located
    on the same node as their parent and would only consider the first
    such successor encountered for each destination.

  * Processing activation messages in remote_dep_get_datatypes,
    remote_dep_mpi_retrieve_datatype would stop iteration on the first
    successor encountered with a valid datatype for that output dep;
    max_priority would be the maximum of all visited successors.

  * parsec_gather_collective_pattern, as part of propagating a
    broadcast, considered the priority of all successor tasks.

These differences meant that max_priority could be different for the
same remote_dep when it was while being sent from the originating node;
when it was being received at a destination; and when it was propagated
as part of a broadcast by an interior node in the broadcast tree.

This commit aligns these diverging behaviors:
  * {parsec|dtd}_release_dep_fct and parsec_gather_collective_pattern
    will consider only successors located on different nodes than the
    parent, but will consider all such successors.
  * The max_priority so determined will be propagated to nodes in the
    broadcast tree by including it in the activation message; processing
    the activation message does not require recomputing the max_priority.

Additionally, correctly reset remote_deps_t in remote_deps_free.