stats: use priorities to determine critical path
Applications often assign task priorities based on the critical path of
the application, if such can be determined statically. We can use this
information to infer a critical path based on the chain of tasks with
highest priority, starting from the initial startup task(s). This allows
us to measure the critical path execution and communication time more
accurately over the entire run of the application, for the subset of
applications that use priorities in this manner.