remote_dep: support priority-based broadcast
Use task priorities to ``rotate'' a broadcast topology around the node
on which the highest-priority successor task is located, so that the
highest-priority successor will receive data and be able to begin
execution sooner. This is not a fully priority-based broadcast: only the
highest-priority successor is prioritized, but this is sufficient for
many applications.

Additionally converts the existing root-based broadcast rotation method
to a configurable compile-time option. These are controlled by the
PARSEC_BCAST_ROTATE CMake cache variable:
  * -DPARSEC_BCAST_ROTATE=OFF (default)
  * -DPARSEC_BCAST_ROTATE=ROOT
  * -DPARSEC_BCAST_ROTATE=PRIORITY