tests/bcast_bug: add new test demonstrating bug #624
Due to how PaRSEC attempts to coalesce broadcasts of overlapping node
sets and how the broadcast trees are constructed, if a node deeper in
the broadcast tree requires data that its parent in the tree did not,
the application may encounter issues (e.g. segfault) when the
nonexistant data is requested from the parent.